+++ src/benchmarks/rxcpp_benchmark.cpp
@@ -799,7 +799,7 @@ TEST_CASE("last")
             });
 
         std::vector<rxcpp::subscriber<int>> subs{};
-        for (size_t i = 0; i < meter.runs(); ++i)
+        for (int i = 0; i < meter.runs(); ++i)
             subs.push_back(rxcpp::make_subscriber<int>());
 
         meter.measure([&](int i)
